# swagger/src/main/resources/swagger-cost-seiko.yaml
openapi: 3.0.3
info:
  title: Cost seiko
  description: |-
    Cost Seiko API
  contact:
    email: khalil@lokalix-solutions.com
  version: 1.0.0
servers:
  - url: http://localhost:9090
...
paths:
...
  /currencies/search:
    post:
      tags:
        - currency
      summary: |
        Search currencies with filter:
        - freetext looking at code
        Return paginated data ordered by code ASC
      operationId: searchCurrencies
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      requestBody:
        description: search currencies
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasicSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CurrenciesPaginated"
  /currencies/{uid}:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - currency
      summary: Retrieve a currency
      operationId: retrieveCurrency
      responses:
        '200':
          description: currency
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
    put:
      tags:
        - currency
      summary: Update a currency
      operationId: updateCurrency
      requestBody:
        description: update currency object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyUpdate'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          $ref: '#/components/responses/BadRequest'
  /currencies:
    get:
      tags:
        - currency
      summary: List all currencies
      operationId: listAllCurrencies
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Currency'
        '400':
          $ref: '#/components/responses/BadRequest'
    post:
      tags:
        - currency
      summary: Create a new currency
      operationId: createCurrency
      requestBody:
        description: Create currency object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CurrencyCreate'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
        '400':
          $ref: '#/components/responses/BadRequest'
  /currencies/{uid}/archive:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - currency
      summary: Archive a currency
      operationId: archiveCurrency
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Currency'
  /currencies/{uid}/history:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - currency
      summary: Get exchange rate history for a currency
      operationId: getCurrencyHistory
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ExchangeRateHistory'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /cost-requests/{uid}/lines/{lineUid}/files/download:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/lineUidParam'
    post:
      tags:
        - cost-request-line
      summary: download a cost request line files
      operationId: downloadCostRequestLineFiles
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Uid"
      responses:
        '200':
          description: Successful operation
...
  /cost-request-lines/to-be-estimated/search:
    post:
      tags:
        - cost-request-line
      summary: |
        Search cost request lines to be estimated with filter
        Return paginated data ordered by creationDate DESC
      operationId: searchCostRequestLinesToBeEstimated
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasicSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomCostRequestLinesPaginated"
  /cost-request-lines/to-be-estimated/{uid}/estimate:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request-line
      summary: estimate a cost request line
      operationId: estimateCostRequestLine
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CostRequestLineEstimate"
        required: true
      responses:
        '200':
          description: successful operation
  /cost-request-lines/to-be-estimated/{uid}/reject:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request-line
      summary: reject a cost request line
      operationId: rejectCostRequestLine
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CostRequestLineReject"
        required: true
      responses:
        '200':
          description: successful operation
  /cost-request-lines/to-be-estimated/{uid}/files:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    get:
      tags:
        - cost-request-line
      summary: get to be estimated cost request line files metadata
      operationId: getToBeEstimatedCostRequestLineFilesMetadata
      parameters:
        - $ref: '#/components/parameters/uidParam'
      responses:
        '200':
          description: successful
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FileInfo'
  /cost-request-lines/to-be-estimated/{uid}/files/download:
    parameters:
      - $ref: '#/components/parameters/uidParam'
    post:
      tags:
        - cost-request-line
      summary: download to be estimated cost request line files
      operationId: downloadToBeEstimatedCostRequestLineFiles
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/Uid"
      responses:
        '200':
          description: Successful operation
  /cost-request-lines/planning/search:
    parameters:
      - $ref: '#/components/parameters/uidParam'
      - $ref: '#/components/parameters/lineUidParam'
    post:
      tags:
        - cost-request-line
      summary: |
        Search cost request lines for the planning
        Return paginated data ordered by creationDate DESC
      operationId: searchCostRequestLinesForPlanning
      parameters:
        - $ref: '#/components/parameters/offsetParam'
        - $ref: '#/components/parameters/limitParam'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BasicSearch'
        required: true
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CustomCostRequestLinesPaginated"
components:
  parameters:
    uidParam:
      in: path
      name: uid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
    offsetParam:
      in: query
      name: offset
      required: true
      schema:
        type: integer
        minimum: 0
        default: 0
      description: The number of items to skip before starting to collect the result set.
    limitParam:
      in: query
      name: limit
      required: true
      schema:
        type: integer
        minimum: 1
        maximum: 50
        default: 20
      description: The numbers of items to return.
...
    lineUidParam:
      in: path
      name: lineUid
      required: true
      schema:
        $ref: '#/components/schemas/Uid'
      description: The line UID
...
  schemas:
...
    CostRequestLineEstimate:
      type: object
      required:
        - unitCostInCurrency
      properties:
        unitCostInCurrency:
          type: number
          format: double
    CostRequestLineReject:
      type: object
      required:
        - rejectReason
      properties:
        rejectReason:
          type: string
...
    PaginationInfos:
      type: object
      required:
        - total
        - page
        - per_page
        - has_next
        - has_prev
      properties:
        total: { type: integer }
        page: { type: integer }
        per_page: { type: integer }
        has_next: { type: boolean }
        has_prev: { type: boolean }
    BasicSearch:
      type: object
      required:
        - searchText
      properties:
        searchText:
          type: string
...
    FileInfo:
      type: object
      required:
        - uid
        - fileName
        - type
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        fileName:
          type: string
        type:
          $ref: '#/components/schemas/FileType'
...
    Uid:
      type: string
      format: uuid
...
    FileType:
      type: string
      enum:
        - ANY
        - QUOTATION_PDF
...
    Customer:
      type: object
      required:
        - uid
        - name
        - code
        - dyson
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        name:
          type: string
        code:
          type: string
        dyson:
          type: boolean
        markup:
          type: number
          format: double
        paymentTerms:
          type: string
        nbShipmentLocations:
          type: integer
...
    ProductName:
      type: object
      required:
        - uid
        - name
        - code
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        name:
          type: string
        code:
          type: string
    CurrenciesPaginated:
      type: object
      allOf:
        - $ref: "#/components/schemas/PaginationInfos"
        - type: object
          required:
            - results
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/Currency"
    CurrencyCreate:
      type: object
      required:
        - code
        - exchangeRates
      properties:
        code:
          type: string
        exchangeRates:
          type: array
          items:
            $ref: '#/components/schemas/ExchangeRateCreate'
    CurrencyUpdate:
      type: object
      required:
        - code
        - exchangeRates
      properties:
        code:
          type: string
        exchangeRates:
          type: array
          items:
            $ref: '#/components/schemas/ExchangeRateUpdate'
    Currency:
      type: object
      required:
        - uid
        - code
        - exchangeRates
        - creationDate
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        code:
          type: string
        exchangeRates:
          type: array
          items:
            $ref: '#/components/schemas/ExchangeRate'
        creationDate:
          type: string
          format: date-time
    ExchangeRate:
      type: object
      required:
        - toCurrencyCode
        - rate
        - creationDate
      properties:
        toCurrencyCode:
          type: string
          description: "Currency code (e.g., USD, EUR, MYR)"
        rate:
          type: number
          format: double
          description: "1 unit of base currency equals <rate> units of target currency"
        creationDate:
          type: string
          format: date-time
    ExchangeRateCreate:
      type: object
      required:
        - toCurrencyId
        - rate
      properties:
        toCurrencyId:
          $ref: '#/components/schemas/Uid'
        rate:
          type: number
          format: double
          description: "1 unit of base currency equals <rate> units of target currency"
    ExchangeRateUpdate:
      type: object
      required:
        - toCurrencyId
        - rate
      properties:
        toCurrencyId:
          $ref: '#/components/schemas/Uid'
        rate:
          type: number
          format: double
          description: "1 unit of base currency equals <rate> units of target currency"
    ExchangeRateHistory:
      type: object
      required:
        - toCurrencyCode
        - oldRate
        - newRate
        - changeDate
      properties:
        toCurrencyCode:
          type: string
        oldRate:
          type: number
          format: double
          description: "Previous exchange rate value"
        newRate:
          type: number
          format: double
          description: "New exchange rate value"
        changeDate:
          type: string
          format: date-time
          description: "Date when the rate was changed"
...
    CustomCostRequestLinesPaginated:
      type: object
      allOf:
        - $ref: "#/components/schemas/PaginationInfos"
        - type: object
          required:
            - results
          properties:
            results:
              type: array
              items:
                $ref: "#/components/schemas/CustomCostRequestLine"
...
    CostRequestLine:
      type: object
      required:
        - uid
        - parentCostRequestUid
        - status
        - customerPartNumber
        - customerPartNumberRevision
        - quantities
        - costingMethodType
        - toolingStrategy
        - creationDate
        - updatedAt
      properties:
        uid:
          $ref: '#/components/schemas/Uid'
        parentCostRequestUid:
          $ref: '#/components/schemas/Uid'
        status:
          $ref: '#/components/schemas/CostRequestStatus'
        customerPartNumber:
          type: string
        customerPartNumberRevision:
          type: string
        description:
          type: string
        productName:
          $ref: '#/components/schemas/ProductName'
        quantities:
          type: array
          items:
            type: integer
        costingMethodType:
          $ref: '#/components/schemas/CostingMethodType'
        toolingStrategy:
          $ref: '#/components/schemas/ToolingStrategy'
        creationDate:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        nbFiles:
          type: integer
        nbMaterialLines:
          type: integer
        nbToolingLines:
          type: integer
        markup:
          type: number
          format: double
        toolingMarkup:
          type: number
          format: double
        priceRejectReason:
          type: string
        outsourced:
          type: boolean
        outsourcingStatus:
          $ref: '#/components/schemas/OutsourcingStatus'
        outsourcingRejectReason:
          type: string
        outsourcedCostInSystemCurrency:
          type: number
          format: double
        procurementStatus:
          $ref: '#/components/schemas/CostRequestLineProcurementStatus'
        nbMessages:
          type: integer
    CustomCostRequestLine:
      allOf:
        - $ref: '#/components/schemas/CostRequestLine'
        - type: object
          required:
            - costRequestReferenceNumber
            - costRequestRevision
            - customer
          properties:
            costRequestReferenceNumber:
              type: string
            costRequestRevision:
              type: integer
            customer:
              $ref: '#/components/schemas/Customer'
...
    CostRequestLineProcurementStatus:
      type: string
      enum:
        - OK
        - PENDING
        - MISSING
    CostingMethodType:
      type: string
      enum:
        - HV
        - BUDGETARY
        - NPI
        - LV
    ToolingStrategy:
      type: string
      enum:
        - AMORTIZED
        - SEPARATED
...
    CostRequestStatus:
      type: string
      enum:
        - PENDING_INFORMATION
        - READY_FOR_REVIEW
        - READY_TO_ESTIMATE
        - READY_TO_VALIDATE
        - READY_FOR_MARKUP
        - PENDING_APPROVAL
        - PRICE_APPROVED
        - PRICE_REJECTED
        - PENDING_REESTIMATION
        - READY_TO_QUOTE
        - ACTIVE
        - WON
        - LOST
        - NEW_REVISION_CREATED
        - ABORTED
...
    OutsourcingStatus:
      type: string
      enum:
        - TO_BE_ESTIMATED
        - ESTIMATED
        - REJECTED
...
    CustomError:
      type: object
      required:
        - code
      properties:
        code:
          $ref: '#/components/schemas/CustomErrorCode'
        message:
          type: string
    CustomErrorCode:
      type: string
      enum:
        - ACCESS_FORBIDDEN
        - INTERNAL_ERROR
        - INVALID_EMAIL_OR_PASSWORD
        - ENTITY_NOT_FOUND
        - ENTITY_ALREADY_EXISTS
        - LICENSE_LIMIT_EXCEEDED
        - USER_IS_INACTIVE
        - USER_RESET_PASSWORD_INVALID_TOKEN
        - USER_CANNOT_DEACTIVATE_ONESELF
        - FILE_UNSUPPORTED
        - FILE_EXISTS
        - FILE_NOT_FOUND
        - GENERIC_ERROR
        - MUST_BE_DISCONNECTED
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CustomError'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes

# 2) Apply the security globally to all operations
security:
  - bearerAuth: [ ]         # use the same name as above